task_5 release, task_7 start
Fixed all memory issues in task_5.

Added a new "memory_manager" prototype for task_7,
assuming the presence of a combination of MT + Intrinsics + MM in the release.

I've been assigned to task_8 and plan to use
Python with CUDA or similar technologies to accomplish it.
It's likely that the solution to this problem will be available
here before task_7.

+++ task_5/src/methods/polynoms.c
@@ -19,7 +19,7 @@ void fft(complex double *x, size_t n, size_t inverse) {
   fft(even, n / 2, inverse);
   fft(odd, n / 2, inverse);
 
-  complex double W_N = cexp((2 * M_PI * I) / n * (inverse ? 1 : -1));
+  complex double W_N = cexp((2 * M_PI * I) / (double)n * (inverse ? 1 : -1));
   complex double W = 1;
   for (size_t k = 0; k < n / 2; k++) {
     complex double t = W * odd[k];
@@ -57,7 +57,7 @@ vector poly_mult(vector *poly_1, vector *poly_2) {
 
   fft(fft_poly_1, n, 1);
   for (size_t i = 0; i < result_deg; ++i) {
-    double tmp = (creal(fft_poly_1[i]) / n);
+    double tmp = (creal(fft_poly_1[i]) / (double)n);
     vector_push(&res, &tmp);
   }
